Rename Configuration class to PluginConfigurationView
Plain "Configuration" would be easily mixed with "settings".
PluginConfigurationView should specify the use case for class much
more precisely.
