+++ src/main/java/com/tw/go/plugin/provider/Provider.java
@@ -1,6 +1,6 @@
 package com.tw.go.plugin.provider;
 
-import com.tw.go.plugin.setting.Configuration;
+import com.tw.go.plugin.setting.PluginConfigurationView;
 import com.tw.go.plugin.setting.PluginSettings;
 
 import java.util.List;
@@ -16,7 +16,7 @@ public interface Provider {
 
     public List<Map<String, Object>> validateConfig(Map<String, Object> fields);
 
-    public Configuration configuration();
+    public PluginConfigurationView configuration();
 
     public PluginSettings pluginSettings(Map<String, String> responseBodyMap);
 }
